requirements.txt
# Core Framework & Server
fastapi==0.119.0
uvicorn[standard]==0.38.0

# Database & ORM
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.0

# Data Models & Settings
pydantic==2.12.3
pydantic-settings==2.11.0

# Task Queue & Monitoring
celery==5.5.3
flower==2.0.1
redis==4.6.0  # ✅ compatible with fastapi-cache2

# Data Science & Analysis
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
pandas-ta>=0.3.14b0
yfinance==0.2.66

# Utility Libraries
python-dotenv>=1.0.0
httpx>=0.25.0
python-jose[cryptography]==3.3.0
structlog>=24.1.0

# API Caching (no redis extra!)
fastapi-cache2==0.2.2  # ✅ installed without extras to prevent duplicate redis deps

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
